% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetFile.R
\name{GetFile}
\alias{GetFile}
\title{Query the Google Analytics Premium API for the specified dimensions, metrics and other query parameters}
\usage{
GetFile(query.builder, token, accountid, webpropertyid, profileid)
}
\arguments{
\item{query.builder}{Name of the object created using \code{\link{QueryBuilder}}}

\item{token}{Name of the token object created using \code{\link{Auth}}}

\item{accountid}{Google analytics premium account id}

\item{webpropertyid}{Webproperty ID for google analytics premium account}

\item{profileid}{View ID for google analytics premium account}
}
\value{
It returns path to file on local drive that contains extracted unsampled data.
}
\description{
Query the Google Analytics Premium API for the specified dimensions, metrics and other query parameters
}

