% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetPostResponse.R
\name{GetPostResponse}
\alias{GetPostResponse}
\title{This will request (POST method) with the prepared Query to the Google Analytics
and returns the response in R object (res).}
\usage{
GetPostResponse(query.uri, postbody, token)
}
\arguments{
\item{query.uri}{The URL to insert report in API.}

\item{postbody}{Body part of POST request.}

\item{token}{Oauth2.0 token.}
}
\value{
res Response of POST request
}
\description{
This will request (POST method) with the prepared Query to the Google Analytics
and returns the response in R object (res).
}
\keyword{internal}

