% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ToBody.R
\name{ToBody}
\alias{ToBody}
\title{Returns the URI constructed from the parameter settings. This also
URI-encodes all the values in each query parameter.}
\usage{
ToBody(query.builder, token)
}
\arguments{
\item{query.builder}{Name of the Object of the Query Builder Class}

\item{token}{Token object containing the OAuth2.0 Authentication details}
}
\value{
A full URI that can be used with the Google Analytics API.
}
\description{
Returns the URI constructed from the parameter settings. This also
URI-encodes all the values in each query parameter.
}

